serverless_backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
APScheduler==3.10.4
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
aws-wsgi==0.2.7
Babel==2.15.0
beam==0.6.0
# beam-sdk==0.15.15
beautifulsoup4==4.12.3
beta9==0.1.47
betterproto-beta9==2.0.0b7
black==24.4.2
bleach==6.1.0
blinker==1.7.0
boto3==1.34.137
botocore==1.34.137
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorhash==2.0.0
comm==0.2.2
configparser==7.0.0
croniter
cryptography==42.0.2
dataclasses-json==0.6.5
debugpy==1.8.2
decorator==4.4.2
deepgram-sdk
defusedxml==0.7.1
distro==1.9.0
durationpy==0.6
executing==2.0.1
elevenlabs==1.7.0
fastapi==0.110.3
fastjsonschema==2.20.0
ffmpeg-python
filelock==3.15.3
firebase-admin==6.4.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-MonitoringDashboard==3.3.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
google-api-core==2.17.0
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
google-cloud-tasks==2.16.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.60.1
grpcio-tools==1.62.2
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
h5df==0.1.5
h5py==3.10.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==5.2.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.13.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kappa==0.6.0
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.55
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow>=3.18.0,<4.0.0
marshmallow_dataclass==8.6.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.30.3
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
placebo==0.9.0
platformdirs==4.2.2
portalocker==2.8.2
proglog==0.1.10
prometheus-flask-exporter==0.23.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytube==15.0.0
pytube3==9.6.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.9.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
scikit-image==0.23.2
scipy==1.13.0
Send2Trash==1.8.3
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tenacity==8.3.0
terminado==0.18.1
text-unidecode==1.3
tifffile==2024.5.10
tiktoken==0.6.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
troposphere==4.8.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==2.0.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.29.0
validators==0.20.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.11
yarl==1.9.4
zappa==0.59.0
zipp==3.19.2
